[ add ] modules to provide instance methods for String
